Fixed output  bug
If the element we are looking for was at the first position it was outputting zero, and remember if the element is not present in the list the location is set to zero. Meaning the value of the location for the element in the first position is the same as for an element which is not present in the list. Hence I fixed the code to give out the location in normal counting which is from 1 going up (1, 2, 3 ...) instead of from 0 going up (0, 1, 2, ...).